requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
asttokens==2.4.1
attrs==24.1.0
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
colorama==0.4.6
comm==0.2.2
crispy-bootstrap5==0.7
cryptography==43.0.0
dataclasses-json==0.6.7
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.7
django-allauth==0.57.0
django-crispy-forms==2.1
django-summernote==0.8.20.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.96
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.38.0
orjson==3.10.6
packaging==24.1
parso==0.8.4
pgvector==0.3.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.5.0
tiktoken==0.7.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==5.3.0
yarl==1.9.4
